Solution - HackthisSite Programming Mission 4
This level is about parsing an randomly generated XML file, available

https://www.hackthissite.org/missions/prog/4/XML <-- (You have to login to get the XML)



The XML file contains information about lines and arcs you have to draw. The result of the drawing gives you the solution to this challenge.
Click here to get more information about the XML file and to see an example. 




Send this:
BlueStringHere,GreenStringHere,RedStringHere,YellowStringHere,WhiteStringHere
as answer.


You have 120 seconds time to send the solution.